+++ src/main/java/dev/nxtime/hidearmor/commands/HideArmorAdminCommand.java
@@ -1,0 +1,44 @@
+package dev.nxtime.hidearmor.commands;
+
+import com.hypixel.hytale.component.Ref;
+import com.hypixel.hytale.component.Store;
+import com.hypixel.hytale.server.core.command.system.CommandContext;
+import com.hypixel.hytale.server.core.command.system.basecommands.AbstractPlayerCommand;
+import com.hypixel.hytale.server.core.entity.entities.Player;
+import com.hypixel.hytale.server.core.universe.PlayerRef;
+import com.hypixel.hytale.server.core.universe.world.storage.EntityStore;
+import com.hypixel.hytale.server.core.universe.world.World;
+
+import javax.annotation.Nonnull;
+
+/**
+ * Admin command to open the default settings configuration GUI.
+ *
+ * @author nxtime
+ * @version 0.6.0
+ */
+public class HideArmorAdminCommand extends AbstractPlayerCommand {
+
+    public HideArmorAdminCommand(String name, String description) {
+        super(name, description);
+    }
+
+    @Override
+    protected void execute(@Nonnull CommandContext context, @Nonnull Store<EntityStore> store,
+            @Nonnull Ref<EntityStore> ref, @Nonnull PlayerRef playerRef, @Nonnull World world) {
+        Player player = store.getComponent(ref, Player.getComponentType());
+        if (player == null)
+            return;
+
+        world.execute(() -> {
+            var playerRefComponent = store.getComponent(ref,
+                    com.hypixel.hytale.server.core.universe.PlayerRef.getComponentType());
+
+            if (playerRefComponent != null) {
+                player.getPageManager().openCustomPage(ref, store,
+                        new dev.nxtime.hidearmor.gui.HideArmorAdminGuiPage(playerRefComponent,
+                                com.hypixel.hytale.protocol.packets.interface_.CustomPageLifetime.CanDismiss));
+            }
+        });
+    }
+}